# api/swagger.yaml
swagger: '2.0'
info:
  description: This is Chiappetta-Gulli-Marabelli e-Commerce Bookstore
  version: 1.0.0
  title: e-Commerce Bookstore
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
host: 'localhost:8080'
basePath: /api
...
schemes:
  - http
paths:
...
  '/user':
    post:
      tags:
        - user
      summary: Adds a new user / registration
      operationId: addUser
      description: Adds a new user / registration
      consumes:
        - application/x-www-form-urlencoded
      produces:
        - application/json
      parameters:
        - name: User
          in: body
          description: a user
          schema:
            $ref: '#/definitions/User'
      responses:
        '201':
          description: item created
        '400':
          description: 'invalid input, object invalid'
        '409':
          description: an existing item already exists
      x-swagger-router-controller: User
  '/user/login':

    post:
      tags:
        - user
      summary: Login
      description: Login with a form
      operationId: userLogin
      consumes:
        - application/x-www-form-urlencoded
      produces:
        - application/json
      parameters:
        - name: email
          in: formData
          required: true
          type: string
        - name: password
          in: formData
          required: true
          type: string
      responses:
        '200':
          description: succesfull login
        '401':
          description: unauthorized
      x-swagger-router-controller: User
  '/user/{ID}':
  
    get:
      tags:
        - user
      summary: Returns the personal info of the user
      description: Returns the personal info of the user which Id is passed
      operationId: getInfo
      produces:
        - application/json
      parameters:
        - name: ID
          in: path
          type: string
          required: true
          description: ID of the user
      responses:
        '200':
          description: Search info
          schema:
            type: object
            items:
              $ref: '#/definitions/User'
        '400':
          description: invalid input
        '404':
          description: user not found
      x-swagger-router-controller: User
      
    put:
      tags:
        - user
      summary: update a user
      operationId: updateUser
      description: Update user checking which field is different
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: ID
          in: path
          type: string
          required: true
          description: ID of the user
        - name: User
          in: body
          description: a User updated
          schema:
            $ref: '#/definitions/User'
      responses:
        '201':
          description: user updated
        '400':
          description: 'invalid input, object invalid'
        '404':
          description: user not found
      x-swagger-router-controller: User
      
    delete:
      tags:
        - user
      summary: delete a user
      operationId: deleteUser
      description: Delete a user from the system
      parameters:
        - name: ID
          in: path
          type: string
          required: true
      responses:
        '201':
          description: item deleted
        '400':
          description: 'invalid input, user doesn''t exist'
        '404':
          description: user not found
    x-swagger-router-controller: User
...
definitions:
...
  User:
    required:
      - email
      - name
      - lastName
    properties:
      ID:
        type: integer
        example: 123
      email:
        type: string
        format: email
        example: pierluigi.sanpietro@mail.polimi.it
      password:
        type: string
        format: password
        example: Th1s-i5_A+zEkoOrE.Pas5w00Rd
      name:
        type: string
        example: Pierluigi
      lastName:
        type: string
        example: Sanpietro
      isAdmin:
        type: boolean
        example: false